requirements.txt
python-telegram-bot[job-queue]
requests
cachetools
colorama
python-dotenv